# AplicacionBodega.API/AplicacionBodega/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AplicacionBodega0, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/StockProducto/test-connection": {
      "get": {
        "tags": [
          "Bodega"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/StockProducto/Add": {
      "post": {
        "tags": [
          "Bodega"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockProducto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/StockProducto/{productoId}": {
      "get": {
        "tags": [
          "Bodega"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StockProducto"
            }
          }
        }
      }
    },
    "/api/StockProducto/Modificar/{productoId}": {
      "put": {
        "tags": [
          "Bodega"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StockProductoUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos/test-connection": {
      "get": {
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos/ObtenerProductos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Productos"
              }
            }
          }
        }
      }
    },
    "/api/Productos/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Productos"
            }
          }
        }
      }
    },
    "/api/Productos/Modificar/{id}": {
      "put": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Productos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos": {
      "post": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Productos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Productos"
            }
          }
        }
      }
    },
    "/api/Productos/Borrar/{id}": {
      "delete": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Productos": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombreProducto": {
          "type": "string"
        },
        "precio": {
          "format": "int32",
          "type": "integer"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "StockProducto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "cantidadComprada": {
          "format": "int32",
          "type": "integer"
        },
        "costoTotal": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "StockProductoUpdateDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "cantidadComprada": {
          "format": "int32",
          "type": "integer"
        },
        "costoTotal": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}